# api/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  description: Configure and inspect a running TDNS server.
  title: TDNS Management API
  version: "1.0"
paths:
...
  "/api/tagger/addr/{tagName}":
    put:
      description: Replace all tags assigned to an address through the legacy endpoint.
      operationId: taggerLegacyAddressReplace
      parameters:
        - description: Address encoded in the legacy path parameter
          in: path
          name: tagName
          required: true
          schema:
            type: string
      requestBody:
        $ref: "#/components/requestBodies/api.ReplaceMemberLabelsRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Replace address tags using the legacy path
      tags:
        - tagger
  /api/tagger/address:
    post:
      description: Set all tags for an address supplied in the request body.
      operationId: taggerAddressCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.MemberLabelsRequest"
        description: Address and tags
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Create or update an address
      tags:
        - tagger
  "/api/tagger/address/{address}":
    put:
      description: Replace all tags assigned to an address.
      operationId: taggerAddressReplace
      parameters:
        - description: IP address or CIDR
          in: path
          name: address
          required: true
          schema:
            type: string
      requestBody:
        $ref: "#/components/requestBodies/api.ReplaceMemberLabelsRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Replace address tags
      tags:
        - tagger
  /api/tagger/hosts:
    get:
      description: Search hosts observed in the DNS log.
      operationId: taggerKnownHostsSearch
      parameters:
        - description: Address or host substring
          in: query
          name: search
          schema:
            type: string
        - description: Maximum result count
          in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Search known hosts
      tags:
        - tagger
  /api/tagger/tags:
    get:
      description: Return all client classification tags.
      operationId: taggerTagsList
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: List tags
      tags:
        - tagger
    post:
      description: Create a client classification tag.
      operationId: taggerTagCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.AddTagRequest"
        description: Tag name
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Create a tag
      tags:
        - tagger
  "/api/tagger/tags/{tagName}":
    delete:
      description: Delete a client classification tag.
      operationId: taggerTagDelete
      parameters:
        - description: Tag name
          in: path
          name: tagName
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Delete a tag
      tags:
        - tagger
    get:
      description: Return addresses assigned to a tag.
      operationId: taggerTagMembersList
      parameters:
        - description: Tag name
          in: path
          name: tagName
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: List tag members
      tags:
        - tagger
    post:
      description: Assign one or more addresses to a tag.
      operationId: taggerTagMembersAdd
      parameters:
        - description: Tag name
          in: path
          name: tagName
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.AddMemberRequest"
        description: Member addresses
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Add tag members
      tags:
        - tagger
  "/api/tagger/tags/{tagName}/{address}":
    delete:
      description: Remove an address from a tag.
      operationId: taggerTagMemberDelete
      parameters:
        - description: Tag name
          in: path
          name: tagName
          required: true
          schema:
            type: string
        - description: IP address or CIDR
          in: path
          name: address
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: string
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                type: string
        "503":
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.Response"
      security:
        - BearerAuth: []
        - CookieAuth: []
      summary: Remove a tag member
      tags:
        - tagger
...
servers:
  - url: /
components:
  requestBodies:
    api.ReplaceMemberLabelsRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/api.ReplaceMemberLabelsRequest"
      description: Replacement tags
      required: true
  securitySchemes:
    BearerAuth:
      description: Bearer token using the format "Bearer {token}".
      in: header
      name: Authorization
      type: apiKey
    CookieAuth:
      description: Opaque HttpOnly browser session cookie set by POST /api/auth/exchange.
      in: cookie
      name: __Host-tdns-session
      type: apiKey
  schemas:
    api.AddMemberRequest:
      properties:
        members:
          items:
            type: string
          type: array
      type: object
    api.AddTagRequest:
      properties:
        name:
          example: trusted
          type: string
      type: object
...
    api.BlacklistStatus:
      properties:
        blockfile_total_entries:
          type: integer
        enabled:
          type: boolean
        excludes:
          items:
            type: string
          type: array
        external_file:
          type: string
        external_pull_period:
          type: string
        external_repo:
          type: string
        external_repo_branch:
          type: string
        file:
          type: string
        persisted_excludes:
          items:
            type: string
          type: array
        persisted_hosts:
          items:
            type: string
          type: array
        runtime_whitelist:
          items:
            type: string
          type: array
      type: object
...
    api.CacheStatus:
      properties:
        enabled:
          type: boolean
        excludes:
          items:
            type: string
          type: array
        hits:
          type: integer
        misses:
          type: integer
        ttl:
          type: integer
      type: object
    api.ClientCandidate:
      properties:
        address:
          example: 192.0.2.10
          type: string
        host:
          example: office
          type: string
      type: object
...
    api.DashboardHourlyPoint:
      properties:
        allowed_queries:
          type: integer
        blocked_queries:
          type: integer
        hour_bucket:
          type: integer
        hour_start:
          type: string
        total_queries:
          type: integer
      type: object
    api.DashboardSummary:
      properties:
        allowed_queries:
          type: integer
        blocked_queries:
          type: integer
        cache_hits:
          type: integer
        cache_misses:
          type: integer
        total_queries:
          type: integer
      type: object
...
    api.HostEntry:
      properties:
        address:
          example: 192.0.2.10
          type: string
        domain:
          example: example.com
          type: string
      type: object
    api.KnownHost:
      properties:
        address:
          example: 192.0.2.10
          type: string
        host:
          example: office
          type: string
      type: object
    api.LogDetails:
      properties:
        counter:
          example: 42
          minimum: 0
          type: integer
        domain:
          example: example.com.
          type: string
        host:
          example: office
          type: string
      type: object
    api.MemberLabelsRequest:
      properties:
        address:
          example: 192.0.2.10
          type: string
        tags:
          items:
            type: string
          type: array
      type: object
    api.ReplaceMemberLabelsRequest:
      properties:
        tags:
          items:
            type: string
          type: array
      type: object
    api.Response:
      properties:
        blacklist:
          $ref: "#/components/schemas/api.BlacklistStatus"
        cache:
          $ref: "#/components/schemas/api.CacheStatus"
        clients:
          items:
            $ref: "#/components/schemas/api.ClientCandidate"
          type: array
        current_status:
          example: enabled
          type: string
        hourly:
          items:
            $ref: "#/components/schemas/api.DashboardHourlyPoint"
          type: array
        items:
          items:
            type: string
          type: array
        kind:
          example: api.tdns/cache/response
          type: string
        known_hosts:
          items:
            $ref: "#/components/schemas/api.KnownHost"
          type: array
        log_items:
          items:
            $ref: "#/components/schemas/api.LogDetails"
          type: array
        message:
          example: Status OK
          type: string
        static_response:
          $ref: "#/components/schemas/api.StaticResponseStatus"
        stub_resolver:
          $ref: "#/components/schemas/api.StubResolverStatus"
        summary:
          $ref: "#/components/schemas/api.DashboardSummary"
        tag_members:
          items:
            $ref: "#/components/schemas/api.TagMember"
          type: array
        window_hours:
          maximum: 336
          minimum: 1
          type: integer
        zen_mode:
          $ref: "#/components/schemas/api.ZenModeStatus"
      type: object
...
    api.StaticResponseStatus:
      properties:
        configured_hosts:
          items:
            $ref: "#/components/schemas/api.HostEntry"
          type: array
        enabled:
          type: boolean
        file:
          type: string
        labels:
          items:
            type: string
          type: array
        persisted_hosts:
          items:
            $ref: "#/components/schemas/api.HostEntry"
          type: array
        runtime_hosts:
          items:
            $ref: "#/components/schemas/api.HostEntry"
          type: array
      type: object
...
    api.StubResolverStatus:
      properties:
        configured_stubs:
          items:
            type: string
          type: array
        enabled:
          type: boolean
        runtime_stubs:
          items:
            type: string
          type: array
      type: object
    api.TagMember:
      properties:
        address:
          example: 192.0.2.10
          type: string
        has_host_alias:
          type: boolean
        host:
          example: office
          type: string
      type: object
...
    api.ZenModeStatus:
      properties:
        configured_domains:
          items:
            type: string
          type: array
        configured_excludes:
          items:
            type: string
          type: array
        duration_minutes:
          type: integer
        enabled:
          type: boolean
        ends_at:
          type: string
        file:
          type: string
        labels:
          items:
            type: string
          type: array
        persisted_domains:
          items:
            type: string
          type: array
        persisted_excludes:
          items:
            type: string
          type: array
        remaining_seconds:
          type: integer
        runtime_domains:
          items:
            type: string
          type: array
        started_at:
          type: string
      type: object
...